Update test to separate channels/without channels test
