+++ example/home/test/test_general.py
@@ -221,26 +221,3 @@ class TestRegisterMutation(BaseGrappleTest):
         # returned after the author has been saved to the database.
         self.assertEqual(data["author"]["name"], self.name)
         self.assertEqual(data["author"]["slug"], self.slug)
-
-
-class TestRegisterSubscription(BaseGrappleTest):
-    def test_subscription(self):
-        query = """
-        {
-          __schema {
-            subscriptionType {
-              fields {
-                name
-              }
-            }
-          }
-        }
-        """
-
-        results = self.client.execute(query)
-        subscriptions = results["data"]["__schema"]["subscriptionType"]["fields"]
-
-        # We check here that the subscription defined in example/home/subscriptions.py and
-        # added in example/home/wagtail_hooks.py is indeed added to the graphene schema.
-        # Note: it is hard to test subscriptions, but there is some place for improvement here.
-        self.assertIn("hello", subscriptions)
