+++ example/requirements.txt
@@ -2,7 +2,7 @@
 black==19.3b0
 
 # Your app requirements.
--r ../requirements-channels.txt
+-r ../requirements.txt
 
 # Your app in editable mode.
 -e ../
